// src/app/api-client/model/workingSessionDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Employe } from './employe';


export interface WorkingSessionDto { 
    id?: number;
    dateSession?: string;
    arrivalTime?: string;
    departureTime?: string;
    durationMinute?: number;
    typeEnregistration?: string;
    employeId?: number;
    employe?: Employe;
}